+++ src/compiler-frontend/src/main/java/org/smoothbuild/compilerfrontend/compile/ast/define/PScope.java
@@ -1,7 +1,6 @@
 package org.smoothbuild.compilerfrontend.compile.ast.define;
 
 import static org.smoothbuild.compilerfrontend.lang.bindings.Bindings.immutableBindings;
-import static org.smoothbuild.compilerfrontend.lang.name.TokenNames.isTypeVarName;
 
 import org.smoothbuild.compilerfrontend.lang.base.Referenceable;
 import org.smoothbuild.compilerfrontend.lang.base.TypeDefinition;
@@ -12,7 +11,6 @@ import org.smoothbuild.compilerfrontend.lang.type.SArrayType;
 import org.smoothbuild.compilerfrontend.lang.type.SFuncType;
 import org.smoothbuild.compilerfrontend.lang.type.SSchema;
 import org.smoothbuild.compilerfrontend.lang.type.SType;
-import org.smoothbuild.compilerfrontend.lang.type.SVar;
 
 public record PScope(
     ImmutableBindings<? extends Referenceable> referencables,
@@ -23,7 +21,7 @@ public record PScope(
 
   public PScope newInnerScope(
       ImmutableBindings<? extends Referenceable> innerReferenceables,
-      ImmutableBindings<PStruct> innerTypes) {
+      ImmutableBindings<PTypeDefinition> innerTypes) {
     return new PScope(
         immutableBindings(referencables, innerReferenceables),
         immutableBindings(types, innerTypes));
@@ -47,9 +45,6 @@ public record PScope(
   }
 
   private SType typeByReference(Fqn fqn) {
-    if (isTypeVarName(fqn.parts().getLast().toString())) {
-      return new SVar(fqn);
-    }
     return types
         .find(fqn)
         .mapOk(TypeDefinition::type)