+++ src/compiler-frontend/src/main/java/org/smoothbuild/compilerfrontend/compile/ast/define/PStruct.java
@@ -5,12 +5,11 @@ import org.smoothbuild.common.base.Strings;
 import org.smoothbuild.common.base.ToStringBuilder;
 import org.smoothbuild.common.log.location.Location;
 import org.smoothbuild.compilerfrontend.lang.base.IdentifiableCode;
-import org.smoothbuild.compilerfrontend.lang.base.TypeDefinition;
 import org.smoothbuild.compilerfrontend.lang.name.Fqn;
 import org.smoothbuild.compilerfrontend.lang.name.NList;
 import org.smoothbuild.compilerfrontend.lang.type.SStructType;
 
-public final class PStruct implements TypeDefinition, PContainer, IdentifiableCode {
+public final class PStruct implements PTypeDefinition, PContainer, IdentifiableCode {
   private final String nameText;
   private final NList<PItem> fields;
   private final Location location;