+++ src/compiler-frontend/src/main/java/org/smoothbuild/compilerfrontend/compile/ast/define/PTypeDefinition.java
@@ -1,0 +1,5 @@
+package org.smoothbuild.compilerfrontend.compile.ast.define;
+
+import org.smoothbuild.compilerfrontend.lang.base.TypeDefinition;
+
+public sealed interface PTypeDefinition extends TypeDefinition permits PStruct, PTypeParam {}