+++ src/compiler-frontend/src/main/java/org/smoothbuild/compilerfrontend/compile/ast/define/PTypeParams.java
@@ -1,0 +1,3 @@
+package org.smoothbuild.compilerfrontend.compile.ast.define;
+
+public sealed interface PTypeParams permits PExplicitTypeParams, PImplicitTypeParams {}