+++ src/compiler-frontend/src/main/java/org/smoothbuild/compilerfrontend/lang/name/TokenNames.java
@@ -1,7 +1,0 @@
-package org.smoothbuild.compilerfrontend.lang.name;
-
-public class TokenNames {
-  public static boolean isTypeVarName(String name) {
-    return !name.isEmpty() && name.chars().allMatch(CharUtils::isUpperCase);
-  }
-}