+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Change Log
 
 Version ?.??.? (????.??.??)
 ---------------------------
+ * changed smooth language: Added explicit type parameters to functions and named values
  * redesigned cli options of `smooth build` command
  * added standard library function `error`
  * added standard library function `filter`